pyproject.toml
[tool.poetry]
name = "jomlah backend"
version = "0.1.0"
description = ""
readme = "README.md"
 package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
django-storages = "^1.14.4"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"
django-cors-headers = "^4.6.0"
channels = "^4.2.0"
celery = "^5.4.0"
django-environ = "^0.11.2"
firebase-admin = "^6.6.0"
daphne = "^4.1.2"
psycopg2-binary = "^2.9.10"
pyotp = "^2.9.0"
django-filter = "^24.3"
twilio = "^9.4.1"
stripe = "^11.4.1"
geopy = "^2.4.1"
apscheduler = "^3.11.0"
whitenoise = "^6.8.2"
gunicorn = "^23.0.0"
redis = "^5.2.1"
cloudinary = "^1.42.1"
django-cloudinary-storage = "^0.3.0"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
